refactor: move drone swapping related stuff into a separate folder
