+++ src/components/map/features/index.ts
@@ -1,0 +1,1 @@
+export { Arrow } from './Arrow';
