+++ src/features/uavs/actions.ts
@@ -3,18 +3,8 @@ import isNil from 'lodash-es/isNil';
 import reject from 'lodash-es/reject';
 import xor from 'lodash-es/xor';
 
-import { adjustMissionMapping } from '~/features/mission/actions';
-import {
-  getMissionMapping,
-  getReverseMissionMapping,
-} from '~/features/mission/selectors';
 import { getSelection } from '~/features/selection/selectors';
 import { setSelection } from '~/features/selection/slice';
-import { SHOW_UPLOAD_JOB } from '~/features/show/constants';
-import {
-  openUploadDialogForJob,
-  putUavsInWaitingQueue,
-} from '~/features/upload/slice';
 import flock from '~/flock';
 import { isUavId, uavIdToGlobalId } from '~/model/identifiers';
 import type { AppThunk } from '~/store/reducers';
@@ -27,57 +17,6 @@ import {
   getUAVIdsWithColorOverride,
 } from './selectors';
 import { _setLEDColorOverride } from './slice';
-import {
-  getSwapAdjustMissionMappingArgs,
-  getSwapPairAffectedUavIds,
-  type SwapApplyPair,
-} from './utils';
-
-export type ApplySwapDronesBatchOptions = {
-  openUploadAfterSwap?: boolean;
-};
-
-export const applySwapDronesBatch =
-  (
-    pairs: SwapApplyPair[],
-    { openUploadAfterSwap = false }: ApplySwapDronesBatchOptions = {}
-  ): AppThunk =>
-  (dispatch, getState) => {
-    const affectedUavIds = new Set<string>();
-
-    for (const pair of pairs) {
-      const state = getState();
-      const reverseMapping = getReverseMissionMapping(state);
-      const mapping = getMissionMapping(state);
-
-      for (const uavId of getSwapPairAffectedUavIds(
-        pair,
-        mapping,
-        reverseMapping
-      )) {
-        affectedUavIds.add(uavId);
-      }
-
-      const args = getSwapAdjustMissionMappingArgs(pair, reverseMapping);
-      if (!args) {
-        continue;
-      }
-
-      dispatch(adjustMissionMapping(args));
-    }
-
-    if (openUploadAfterSwap && affectedUavIds.size > 0) {
-      const mapping = getMissionMapping(getState());
-      const uploadTargets = [...affectedUavIds].filter((uavId) =>
-        mapping.includes(uavId)
-      );
-
-      if (uploadTargets.length > 0) {
-        dispatch(openUploadDialogForJob({ job: SHOW_UPLOAD_JOB }));
-        dispatch(putUavsInWaitingQueue(uploadTargets));
-      }
-    }
-  };
 
 /**
  * Clears any color override related to the given UAV.
