+++ src/features/uavs/constants.ts
@@ -40,15 +40,3 @@ export const UAV_TOOLBAR_MULTI_SELECTION_LIMIT = 100;
  * Timout for the async compass calibration operation.
  */
 export const COMPASS_CALIB_TIMEOUT = 90;
-
-/**
- * Width of the drone selector column of the swap drones dialog. Kept constant so
- * that showing or hiding the pending swap queue does not reflow the form.
- */
-export const SWAP_DRONES_FORM_COLUMN_WIDTH = 550;
-
-/**
- * Width of the pending swap queue column of the swap drones dialog, including its
- * separator and the spacing around it.
- */
-export const SWAP_DRONES_QUEUE_COLUMN_WIDTH = 200;
