+++ src/features/uavs/details.ts
@@ -1,5 +1,6 @@
 /**
- * @file Redux slices for UAV-related dialogs registered under `state.dialogs`.
+ * @file Slice of the state object that handles the state of the UAV details
+ * dialog.
  */
 
 import { createSlice, type PayloadAction } from '@reduxjs/toolkit';
@@ -104,28 +105,5 @@ export const getUAVDetailsDialogPosition: AppSelector<Coordinate2DObject> = (
 
 export const getUAVDetailsDialogWidth: AppSelector<number> = (state) =>
   state.dialogs.uavDetails.width;
-
-const { actions: swapDronesDialogActions, reducer: swapDronesDialogReducer } =
-  createSlice({
-    name: 'swap-drones-dialog',
-    initialState: { open: false },
-    reducers: {
-      closeSwapDronesDialog: noPayload<{ open: boolean }>((state) => {
-        state.open = false;
-      }),
-
-      showSwapDronesDialog: noPayload<{ open: boolean }>((state) => {
-        state.open = true;
-      }),
-    },
-  });
-
-export const { closeSwapDronesDialog, showSwapDronesDialog } =
-  swapDronesDialogActions;
-
-export const isSwapDronesDialogOpen: AppSelector<boolean> = (state) =>
-  state.dialogs.swapDrones.open;
-
-export { swapDronesDialogReducer };
 
 export default reducer;
