+++ src/features/uavs/swapping/SwapDroneBadge.tsx
