+++ src/features/uavs/swapping/SwapDroneField.tsx
