+++ src/features/uavs/swapping/SwapDronesConfirmDialog.tsx
