+++ src/features/uavs/swapping/SwapDronesPreview.tsx
