+++ src/features/uavs/swapping/SwapDronesQueuePanel.tsx
