+++ src/features/uavs/swapping/selectors.ts
@@ -1,0 +1,4 @@
+import type { AppSelector } from '~/store/reducers';
+
+export const isSwapDronesDialogOpen: AppSelector<boolean> = (state) =>
+  state.dialogs.swapDrones.open;
