+++ src/features/uavs/swapping/utils.ts
