+++ src/store/reducers.ts
@@ -32,9 +32,8 @@ import showConfiguratorReducer, {
   type ShowData,
 } from '~/features/show-configurator/slice';
 import tasksReducer from '~/features/tasks/slice';
-import uavDetailsDialogReducer, {
-  swapDronesDialogReducer,
-} from '~/features/uavs/details';
+import uavDetailsDialogReducer from '~/features/uavs/details';
+import swapDronesDialogReducer from '~/features/uavs/swapping/slice';
 
 /**
  * Reducer functions for handling the top-level keys of the state object.
